+++ test/metadata-registry/metadataResolver.test.ts
@@ -15,13 +15,14 @@ import { nls } from '../../src/i18n';
 import {
   mockRegistry,
   xmlInFolder,
-  keanu,
+  matchingContentFile,
   taraji,
   mixedContentInFolder,
   simon,
   sean,
   gene,
   mockRegistryData,
+  decomposedtoplevel,
 } from '../mock/registry';
 import { join, basename, dirname } from 'path';
 import { TypeInferenceError } from '../../src/errors';
@@ -43,6 +44,7 @@ import {
   TARAJI_VIRTUAL_FS,
   TARAJI_XML_PATHS,
 } from '../mock/registry/tarajiConstants';
+import { ComponentSet } from '../../src';
 
 const testUtil = new RegistryTestUtil();
 
@@ -54,7 +56,7 @@ describe('MetadataResolver', () => {
 
     describe('File Paths', () => {
       it('Should throw file not found error if given path does not exist', () => {
-        const path = keanu.KEANU_SOURCE_PATHS[0];
+        const path = matchingContentFile.CONTENT_PATHS[0];
 
         assert.throws(
           () => resolver.getComponentsFromPath(path),
@@ -64,42 +66,42 @@ describe('MetadataResolver', () => {
       });
 
       it('Should determine type for metadata file with known suffix', () => {
-        const path = keanu.KEANU_XML_PATHS[0];
+        const path = matchingContentFile.XML_PATHS[0];
         const access = testUtil.createMetadataResolver([
           {
-            dirPath: keanu.KEANUS_DIR,
-            children: [keanu.KEANU_SOURCE_NAMES[0], keanu.KEANU_XML_NAMES[0]],
+            dirPath: matchingContentFile.TYPE_DIRECTORY,
+            children: [matchingContentFile.CONTENT_NAMES[0], matchingContentFile.XML_NAMES[0]],
           },
         ]);
         testUtil.stubAdapters([
           {
-            type: mockRegistryData.types.keanureeves,
+            type: mockRegistryData.types.matchingcontentfile,
             componentMappings: [
               {
                 path,
-                component: keanu.KEANU_COMPONENT,
+                component: matchingContentFile.COMPONENT,
               },
             ],
           },
         ]);
-        expect(access.getComponentsFromPath(path)).to.deep.equal([keanu.KEANU_COMPONENT]);
+        expect(access.getComponentsFromPath(path)).to.deep.equal([matchingContentFile.COMPONENT]);
       });
 
       it('Should determine type for source file with known suffix', () => {
-        const path = keanu.KEANU_SOURCE_PATHS[0];
+        const path = matchingContentFile.CONTENT_PATHS[0];
         const access = testUtil.createMetadataResolver([
           {
-            dirPath: keanu.KEANUS_DIR,
-            children: [keanu.KEANU_SOURCE_NAMES[0], keanu.KEANU_XML_NAMES[0]],
+            dirPath: matchingContentFile.TYPE_DIRECTORY,
+            children: [matchingContentFile.CONTENT_NAMES[0], matchingContentFile.XML_NAMES[0]],
           },
         ]);
         testUtil.stubAdapters([
           {
-            type: mockRegistryData.types.keanureeves,
-            componentMappings: [{ path, component: keanu.KEANU_COMPONENT }],
+            type: mockRegistryData.types.matchingcontentfile,
+            componentMappings: [{ path, component: matchingContentFile.COMPONENT }],
           },
         ]);
-        expect(access.getComponentsFromPath(path)).to.deep.equal([keanu.KEANU_COMPONENT]);
+        expect(access.getComponentsFromPath(path)).to.deep.equal([matchingContentFile.COMPONENT]);
       });
 
       it('Should determine type for path of mixed content type', () => {
@@ -120,21 +122,23 @@ describe('MetadataResolver', () => {
       });
 
       it('Should determine type for path content files', () => {
-        const path = keanu.KEANU_SOURCE_PATHS[0];
+        const path = matchingContentFile.CONTENT_PATHS[0];
         const access = testUtil.createMetadataResolver([
           {
             dirPath: dirname(path),
-            children: keanu.KEANU_SOURCE_NAMES,
+            children: matchingContentFile.CONTENT_NAMES,
           },
         ]);
         testUtil.stubAdapters([
           {
-            type: mockRegistryData.types.keanureeves,
-            componentMappings: [{ path, component: keanu.KEANU_CONTENT_COMPONENT }],
+            type: mockRegistryData.types.matchingcontentfile,
+            componentMappings: [{ path, component: matchingContentFile.CONTENT_COMPONENT }],
             allowContent: false,
           },
         ]);
-        expect(access.getComponentsFromPath(path)).to.deep.equal([keanu.KEANU_CONTENT_COMPONENT]);
+        expect(access.getComponentsFromPath(path)).to.deep.equal([
+          matchingContentFile.CONTENT_COMPONENT,
+        ]);
       });
 
       it('Should determine type for inFolder path content files', () => {
@@ -215,7 +219,7 @@ describe('MetadataResolver', () => {
       });
 
       it('Should not return a component if path to metadata xml is forceignored', () => {
-        const path = keanu.KEANU_XML_PATHS[0];
+        const path = matchingContentFile.XML_PATHS[0];
         const access = testUtil.createMetadataResolver([
           {
             dirPath: dirname(path),
@@ -225,16 +229,16 @@ describe('MetadataResolver', () => {
         testUtil.stubForceIgnore({ seed: path, deny: [path] });
         testUtil.stubAdapters([
           {
-            type: mockRegistryData.types.keanureeves,
+            type: mockRegistryData.types.matchingcontentfile,
             // should not be returned
-            componentMappings: [{ path, component: keanu.KEANU_COMPONENT }],
+            componentMappings: [{ path, component: matchingContentFile.COMPONENT }],
           },
         ]);
         expect(access.getComponentsFromPath(path).length).to.equal(0);
       });
 
       it('Should not return a component if path to content metadata xml is forceignored', () => {
-        const path = keanu.KEANU_XML_PATHS[0];
+        const path = matchingContentFile.XML_PATHS[0];
         const access = testUtil.createMetadataResolver([
           {
             dirPath: dirname(path),
@@ -244,9 +248,9 @@ describe('MetadataResolver', () => {
         testUtil.stubForceIgnore({ seed: path, deny: [path] });
         testUtil.stubAdapters([
           {
-            type: mockRegistryData.types.keanureeves,
+            type: mockRegistryData.types.matchingcontentfile,
             // should not be returned
-            componentMappings: [{ path, component: keanu.KEANU_COMPONENT }],
+            componentMappings: [{ path, component: matchingContentFile.COMPONENT }],
           },
         ]);
         expect(access.getComponentsFromPath(path).length).to.equal(0);
@@ -298,20 +302,20 @@ describe('MetadataResolver', () => {
       });
 
       it('Should walk all file and directory children', () => {
-        const { KEANUS_DIR } = keanu;
-        const stuffDir = join(KEANUS_DIR, 'hasStuff');
-        const noStuffDir = join(KEANUS_DIR, 'noStuff');
-        const kathyXml = join(KEANUS_DIR, xmlInFolder.XML_NAMES[0]);
-        const keanuXml = keanu.KEANU_XML_PATHS[0];
-        const keanuSrc = keanu.KEANU_SOURCE_PATHS[0];
-        const keanuXml2 = join(stuffDir, keanu.KEANU_XML_NAMES[1]);
-        const keanuSrc2 = join(stuffDir, keanu.KEANU_SOURCE_NAMES[1]);
+        const { TYPE_DIRECTORY: MCF_DIR } = matchingContentFile;
+        const stuffDir = join(MCF_DIR, 'hasStuff');
+        const noStuffDir = join(MCF_DIR, 'noStuff');
+        const kathyXml = join(MCF_DIR, xmlInFolder.XML_NAMES[0]);
+        const mcfXml = matchingContentFile.XML_PATHS[0];
+        const mcfContent = matchingContentFile.CONTENT_PATHS[0];
+        const mcfXml2 = join(stuffDir, matchingContentFile.XML_NAMES[1]);
+        const mcfContent2 = join(stuffDir, matchingContentFile.CONTENT_NAMES[1]);
         const tree = new VirtualTreeContainer([
           {
-            dirPath: KEANUS_DIR,
+            dirPath: MCF_DIR,
             children: [
-              basename(keanuXml),
-              basename(keanuSrc),
+              basename(mcfXml),
+              basename(mcfContent),
               xmlInFolder.XML_NAMES[0],
               'hasStuff',
               'noStuff',
@@ -323,15 +327,15 @@ describe('MetadataResolver', () => {
           },
           {
             dirPath: stuffDir,
-            children: [basename(keanuSrc2), basename(keanuXml2)],
+            children: [basename(mcfContent2), basename(mcfXml2)],
           },
         ]);
-        const keanuComponent2: SourceComponent = new SourceComponent(
+        const mcfComponent2: SourceComponent = new SourceComponent(
           {
             name: 'b',
-            type: mockRegistryData.types.keanureeves,
-            xml: keanuXml2,
-            content: keanuSrc2,
+            type: mockRegistryData.types.matchingcontentfile,
+            xml: mcfXml2,
+            content: mcfContent2,
           },
           tree
         );
@@ -349,29 +353,29 @@ describe('MetadataResolver', () => {
             type: mockRegistryData.types.xmlinfolder,
             componentMappings: [
               {
-                path: join(KEANUS_DIR, xmlInFolder.XML_NAMES[0]),
+                path: join(MCF_DIR, xmlInFolder.XML_NAMES[0]),
                 component: kathyComponent2,
               },
             ],
           },
           {
-            type: mockRegistryData.types.keanureeves,
+            type: mockRegistryData.types.matchingcontentfile,
             componentMappings: [
               {
-                path: keanuXml,
-                component: keanu.KEANU_COMPONENT,
+                path: mcfXml,
+                component: matchingContentFile.COMPONENT,
               },
               {
-                path: keanuXml2,
-                component: keanuComponent2,
+                path: mcfXml2,
+                component: mcfComponent2,
               },
             ],
           },
         ]);
-        expect(access.getComponentsFromPath(KEANUS_DIR)).to.deep.equal([
-          keanu.KEANU_COMPONENT,
+        expect(access.getComponentsFromPath(MCF_DIR)).to.deep.equal([
+          matchingContentFile.COMPONENT,
           kathyComponent2,
-          keanuComponent2,
+          mcfComponent2,
         ]);
       });
 
@@ -504,7 +508,7 @@ describe('MetadataResolver', () => {
        * Because files of a mixed content type could have any suffix, they might collide
        * with a type that uses the "suffix index" in the registry and be assigned the incorrect type.
        *
-       * Pretend that this bundle's root xml suffix is the same as KeanuReeves - still should be
+       * Pretend that this bundle's root xml suffix is the same as MatchingContentFile - still should be
        * identified as SimonPegg type
        */
       it('should handle suffix collision for mixed content types', () => {
@@ -515,7 +519,7 @@ describe('MetadataResolver', () => {
           },
           {
             dirPath: simon.SIMON_BUNDLE_PATH,
-            children: [keanu.KEANU_XML_NAMES[0], basename(simon.SIMON_SOURCE_PATHS[0])],
+            children: [matchingContentFile.XML_NAMES[0], basename(simon.SIMON_SOURCE_PATHS[0])],
           },
         ]);
         const access = new MetadataResolver(mockRegistry, tree);
@@ -524,7 +528,7 @@ describe('MetadataResolver', () => {
             {
               name: 'a',
               type: mockRegistryData.types.simonpegg,
-              xml: join(simon.SIMON_BUNDLE_PATH, keanu.KEANU_XML_NAMES[0]),
+              xml: join(simon.SIMON_BUNDLE_PATH, matchingContentFile.XML_NAMES[0]),
               content: simon.SIMON_BUNDLE_PATH,
             },
             tree
@@ -557,6 +561,140 @@ describe('MetadataResolver', () => {
           },
         ]);
         expect(access.getComponentsFromPath(dirPath).length).to.equal(0);
+      });
+    });
+
+    describe('Filtering', () => {
+      it('should only return components present in filter', async () => {
+        const resolver = testUtil.createMetadataResolver([
+          {
+            dirPath: xmlInFolder.COMPONENT_FOLDER_PATH,
+            children: xmlInFolder.XML_NAMES,
+          },
+        ]);
+        const componentMappings = xmlInFolder.XML_PATHS.map((p: string, i: number) => ({
+          path: p,
+          component: xmlInFolder.COMPONENTS[i],
+        }));
+        testUtil.stubAdapters([
+          {
+            type: mockRegistryData.types.xmlinfolder,
+            componentMappings,
+          },
+        ]);
+        const toFilter = {
+          fullName: xmlInFolder.COMPONENTS[0].fullName,
+          type: mockRegistryData.types.xmlinfolder,
+        };
+        const filter = new ComponentSet([toFilter]);
+
+        const result = resolver.getComponentsFromPath(xmlInFolder.COMPONENT_FOLDER_PATH, filter);
+
+        expect(result).to.deep.equal([xmlInFolder.COMPONENTS[0]]);
+      });
+
+      it('should resolve child components when present in filter', async () => {
+        const resolver = testUtil.createMetadataResolver(decomposedtoplevel.DECOMPOSED_VIRTUAL_FS);
+        const children = decomposedtoplevel.DECOMPOSED_TOP_LEVEL_COMPONENT.getChildren();
+        const componentMappings = children.map((c: SourceComponent) => ({
+          path: c.xml,
+          component: c,
+        }));
+        componentMappings.push({
+          path: decomposedtoplevel.DECOMPOSED_TOP_LEVEL_XML_PATH,
+          component: decomposedtoplevel.DECOMPOSED_TOP_LEVEL_COMPONENT,
+        });
+        testUtil.stubAdapters([
+          {
+            type: mockRegistryData.types.decomposedtoplevel,
+            componentMappings,
+          },
+        ]);
+        const toFilter = [
+          {
+            fullName: children[0].fullName,
+            type: children[0].type,
+          },
+          {
+            fullName: children[1].fullName,
+            type: children[1].type,
+          },
+        ];
+        const filter = new ComponentSet(toFilter);
+
+        const result = resolver.getComponentsFromPath(
+          decomposedtoplevel.DECOMPOSED_TOP_LEVEL_COMPONENT_PATH,
+          filter
+        );
+
+        expect(result).to.deep.equal(children);
+      });
+
+      it('should resolve directory component if in filter', () => {
+        const resolver = testUtil.createMetadataResolver([
+          {
+            dirPath: simon.SIMON_DIR,
+            children: [basename(simon.SIMON_BUNDLE_PATH)],
+          },
+          {
+            dirPath: simon.SIMON_BUNDLE_PATH,
+            children: simon.SIMON_SOURCE_PATHS.map((p) => basename(p)).concat([
+              simon.SIMON_XML_NAME,
+            ]),
+          },
+        ]);
+        testUtil.stubAdapters([
+          {
+            type: mockRegistryData.types.simonpegg,
+            componentMappings: [
+              {
+                path: simon.SIMON_BUNDLE_PATH,
+                component: simon.SIMON_COMPONENT,
+              },
+            ],
+          },
+        ]);
+        const filter = new ComponentSet([
+          {
+            fullName: simon.SIMON_COMPONENT.fullName,
+            type: simon.SIMON_COMPONENT.type,
+          },
+        ]);
+
+        const result = resolver.getComponentsFromPath(simon.SIMON_DIR, filter);
+
+        expect(result).to.deep.equal([simon.SIMON_COMPONENT]);
+      });
+
+      it('should not resolve directory component if not in filter', () => {
+        const resolver = testUtil.createMetadataResolver([
+          {
+            dirPath: simon.SIMON_DIR,
+            children: [basename(simon.SIMON_BUNDLE_PATH)],
+          },
+          {
+            dirPath: simon.SIMON_BUNDLE_PATH,
+            children: simon.SIMON_SOURCE_PATHS.map((p) => basename(p)).concat([
+              simon.SIMON_XML_NAME,
+            ]),
+          },
+        ]);
+        testUtil.stubAdapters([
+          {
+            type: mockRegistryData.types.simonpegg,
+            componentMappings: [
+              {
+                path: simon.SIMON_BUNDLE_PATH,
+                component: simon.SIMON_COMPONENT,
+              },
+            ],
+          },
+        ]);
+        const filter = new ComponentSet();
+
+        const result = resolver.getComponentsFromPath(simon.SIMON_DIR, filter);
+
+        expect(result).to.deep.equal([]);
       });
     });
   });